chore(avante): set auto_suggestion option to false
avante.nvim has a hidden keybinding that I haven't found yet that I
activated on accident once and couldn't deactivate that turns on auto
suggestions. Needless to say that using Claude 3.7 for live suggestions
is a bad idea. Not thank you to avante for that (on top of that the
suggestion system was broken so I wasn't really getting any suggestions
it was just consuming a LOT of tokens for nothing)